turnstile, 게시글 관련 프로시저 추가

+++ demo/src/main/java/com/example/demo/controller/todo/ToDoController.java
@@ -39,7 +39,7 @@ public class ToDoController {
 
     // ToDo 삭제 컨트롤러
     @DeleteMapping("/todo/delete/{idx}")
-    public ResponseEntity<?> deleteToDoList(@PathVariable("idx") Integer idx) {
+    public ResponseEntity<?> deleteToDoList(@PathVariable("idx") Long idx) {
         ts.deleteToDoLis(idx);
         return ResponseEntity.ok("삭제 완료");
     }
